Find workaround solution for jQuery init
This preamble must be contained in every file which assumes
a global jQuery object.
Previously, init of the Markdown JS parser failed due to a
missing jQuery global. This is due to a lack of ordering
of JS initialization: the jQuery init from django-markdown
and the admin interface were competing. In my setup, this
lead to following script load order:

1. django-markdown jquery.init.js tries to populate a
   global jQuery object.
2. django-admin jquery.init.js hides jQuery behind
   django.jQuery and restores the global variables to
   an a-priori seen value: undefined.
3. jQuery.markitup.js fails, because global jQuery is
   undefined.

This patch makes jQuery.markitup.js 'see' the jQuery
instance behind django.jQuery.